@@ -18,21 +18,23 @@ namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing
 
         public IEnumerable<PluralRule> Parse()
         {
-            var root = _rulesDocument.DocumentElement;
-
+            var root = _rulesDocument.DocumentElement!;
+            
             foreach(XmlNode dataElement in root.ChildNodes)
             {
-                if(dataElement.Name == "plurals")
+                if (dataElement.Name != "plurals")
                 {
-                    foreach (XmlNode rule in dataElement.ChildNodes)
+                    continue;
+                }
+                
+                foreach (XmlNode rule in dataElement.ChildNodes)
+                {
+                    if(rule.Name == "pluralRules")
                     {
-                        if(rule.Name == "pluralRules")
+                        var parsed = ParseSingleRule(rule);
+                        if (parsed != null)
                         {
-                            var parsed = ParseSingleRule(rule);
-                            if (parsed != null)
-                            {
-                                yield return parsed;
-                            }
+                            yield return parsed;
                         }
                     }
                 }
@@ -41,7 +43,7 @@ namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing
 
         private PluralRule? ParseSingleRule(XmlNode rule)
         {
-            var locales = rule.Attributes["locales"].Value.Split(' ');
+            var locales = rule.Attributes!["locales"].Value.Split(' ');
 
             if (locales.All(l => _excludedLocales.Contains(l)))
             {
@@ -53,7 +55,7 @@ namespace Jeffijoe.MessageFormat.MetadataGenerator.Plural.Parsing
             {
                 if (condition.Name == "pluralRule")
                 {
-                    var count = condition.Attributes["count"].Value;
+                    var count = condition.Attributes!["count"].Value;
 
                     // Ignore other, because other is basically everything else except for the conditions present
                     if (count == "other")
