@@ -2,6 +2,6 @@
 {
     internal static partial class PluralRulesMetadata
     {
-        public static partial bool TryGetRuleByLocale(string locale, out ContextPluralizer pluralizer);
+        public static partial bool TryGetRuleByLocale(string locale, out ContextPluralizer contextPluralizer);
     }
 }
