@@ -1,5 +1,4 @@
-﻿using System;
-using System.Text;
+﻿using System.Text;
 using Microsoft.Extensions.ObjectPool;
 
 namespace Jeffijoe.MessageFormat
